requirements.txt
Django==4.2.4
requests==2.31.0
django-debug-toolbar==4.2.0
psycopg2-binary==2.9.7
mysqlclient==2.2.0
python-dotenv==1.0.0
python-telegram-bot==13.7